pyproject.toml
[project]
name = "golf-fiddling"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.0",
    "pillow>=12.0.0",
    "pygame>=2.6.1",
]

[project.scripts]
golf-editor = "editor.main:main"
golf-dump = "tools.dump:main"
golf-analyze = "tools.analyze:main"
golf-visualize = "tools.visualize:main"
golf-hex2bin = "tools.hex2bin:main"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["golf", "editor", "tools"]
